Planner: extract select module; slim mod.rs to public API
